[api] Reorganize handlers
